+++ src/services/user/services/AdminUsers.js
@@ -88,6 +88,14 @@ class AdminUsers {
 			if (clientQuery.createdAt) query.createdAt = clientQuery.createdAt;
 			if (clientQuery.firstName) query.firstName = clientQuery.firstName;
 			if (clientQuery.lastName) query.lastName = clientQuery.lastName;
+			if (clientQuery.usersForConsent) query._id = clientQuery.usersForConsent;
+			if (clientQuery.searchQuery) {
+				query.$or = [
+					{ firstName: { $regex: clientQuery.searchQuery, $options: 'i' } },
+					{ lastName: { $regex: clientQuery.searchQuery, $options: 'i' } },
+					{ email: { $regex: clientQuery.searchQuery, $options: 'i' } },
+				];
+			}
 
 			return new Promise((resolve, reject) => userModel.aggregate(createMultiDocumentAggregation(query)).option({
 				collation: { locale: 'de', caseLevel: true },
@@ -171,6 +179,7 @@ class AdminUsers {
 
 const formatBirthdayOfUsers = ({ result: { data: users } }) => {
 	users.forEach((user) => {
+		if (user.birthday) user.birthday = moment(user.birthday).format('DD.MM.YYYY');
 		if (user.birthday) {
 			user.birthday = moment(user.birthday).format('DD.MM.YYYY');
 		}