+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `Security`
 
 ## Unreleased
+
+### Added
+- SC-5612 - Adding search feature to the admintables for nuxt-client.
+
 ## [24.4.1] - 2020-08-31
 
 ### Fixed
@@ -19,7 +23,6 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 - SC-5987 Internationalisation: extend user and school model with default language
 
 ### Added
-
 - SC-6172: added hooks and checks to look for unique and not disposable emails in adminUsers service
 
 ## [24.4.0] - 2020-8-31